refactor(tests): split helpers.ts into domain modules
- Split the 672-line tests/helpers.ts into 8 domain modules under
  tests/helpers/ (paths, cli, git, platform, sandbox, templates,
  command-specs, esm), each under 200 lines
- Keep tests/helpers.ts as a barrel re-export (26 values + 3 types)
  so all 30 caller test files stay unchanged
- Adjust rootDir to "../.." for the new module depth; npm test green
  (533 passed / 1 skipped)
